flytt endepunkt og state for personer med ugyldig etterbetalingsperiode inn i hook
